refactor: migrate larger components over to ts
